chore: extract period init helpers into shared module
Moves initializePeriodType/getDefaultPeriods/initializePeriods out of
initializeThematicLayer.js into components/edit/shared/initializePeriod.js
so the Events dialog can reuse them for its own period tab, without
duplicating logic. Pure extraction, no behavior change.
